+++ remix/package.json
@@ -14,13 +14,17 @@
     "@remix-run/node": "^1.4.1",
     "@remix-run/react": "^1.4.1",
     "@remix-run/serve": "^1.4.1",
+    "jsdom": "^20.0.0",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "tape": "^5.5.3",
+    "uvu": "^0.5.3"
   },
   "devDependencies": {
     "@remix-run/dev": "^1.4.1",
     "@remix-run/eslint-config": "^1.4.1",
-    "eslint": "^8.11.0"
+    "eslint": "^8.11.0",
+    "jest": "^28.1.1"
   },
   "engines": {
     "node": ">=14"
